Changes on index.js file and app.js file
